Improve pywbem exception handling (#75)
This PR improves exception handling of the `pywbem` module, which uses different exception calls before and starting with pywbem 1.0.0. 

This allows correct exception handling with older `pywbem` versions (< 1.0.0) and newer versions.

The PR also adds exception handling for HTTP exception, when a HTTP server responds on the requested host and port, but it's not an ESXi CIM server.